# Improve V base64 memory usage by using pre allocated buffers.