Before this PR, on my PC:
```shell
0[13:21:51] /v/benchmarks/base64 LINUX $ ###BEFORE
0[13:22:00] /v/benchmarks/base64 LINUX $ /v/nv/v -prod -cc gcc -o base64_v_gcc test.v
0[13:22:08] /v/benchmarks/base64 LINUX $ ../xtime.rb  ./base64_v_gcc
encode aaaa... to YWFh...:   1464 ms | 1431666688
decode YWFh... to aaaa...:   2278 ms | 1073741824
3.84s, 2397.9Mb
0[13:22:14] /v/benchmarks/base64 LINUX $
```

After this PR, on my PC:
```shell
0[13:22:25] /v/benchmarks/base64 LINUX $ ###AFTER
0[13:22:35] /v/benchmarks/base64 LINUX $ /v/nv/v -prod -cc gcc -o base64_v_gcc test.v
0[13:22:40] /v/benchmarks/base64 LINUX $ ../xtime.rb  ./base64_v_gcc
encode aaaa... to YWFh...:   1095 ms | 1431666688
decode YWFh... to aaaa...:   2009 ms | decode: 1073741824
  3104 ms | <=== total time spent running base64 encode/decode ops
 ok, fail, total =     2,     0,     2
3.11s, 2.0Mb
0[13:22:46] /v/benchmarks/base64 LINUX $
```