@kostya 2019-10-29
i not sure this is correct, comparing to others, all implementations use heap allocations (except c which use stack allocation).