@nuald 2019-10-29
@kostya is right, however it's obvious that the test has the problem with the memory allocation. I've looked into the generated C code and yes, the allocated buffer is not freed, therefore it has the performance problem (as `malloc` can't reuse the previously allocated memory with `sbrk` as we didn't free it). I've created a PR #196 to address it and it definitely shows the performance improvements.