@spytheman 2019-11-07
As it is, the v test measure the speed of malloc/free more than it measures actual base64 encoding/decoding.

See commit #befeaf855d59e6ecde9fa3de264ccbd3d1b48342 , where one of the main optimizations for multiple languages was making the hot loops use a preallocated buffer in one way or another, and eliminating calling malloc/free inside the loops.

Calling manually `free` fixes a memory leak with the current language implementation, but calling base64.decode_in_buffer instead of base64.decode and base64.encode_in_buffer instead of base64.encode , does the same as befeaf855 .