@spytheman 2019-11-07
Should I make another PR, that replaces calls to base64.encode with base64.encode_in_buffer ?