@nuald 2019-11-07
@spytheman `malloc`/`free` on small sizes are virtually free as they use `sbrk` (so it should reuse the same memory, just move the pointer). If you measure the performance of the current version and the version using `base64.encode_in_buffer` you won't see any difference.

However the latter one would be unfair in the general use-case as all other tests allocates the memory in the loop (as we measure "encoding/decoding into a new buffer" as it's not possible to measure just "encoding/decoding" for high-level languages as they don't support pre-allocated buffers). However, if V would support allocating on stack, it would be nice as in that case it would be possible to have closer-to-C code.