@spytheman 2020-01-12
@nuald , V does support allocating on the stack, just do: `buf := [1000]byte` and you will get stack allocated 10000 bytes. In this regard, v is closer to C and should have similar performance.