@nuald 2020-01-14
Looks like the stack allocation doesn't work with the variable sizes, I'm getting compilation errors like:
```
test.v:30:34: undefined: `byte`
   28|            str2_size := 4 * ((str_size + 2) / 3)
   29|            str2 := [str2_size]byte
                                        ^
```
No errors if use numeric values.

Please note though that the memory manipulations barely change anything here - it's not the main culprit here. aklomp uses SSE instructions, pure C version uses some optimizations too like `__builtin_bswap32`, the V base64 module doesn't have any optimizations though. It's still the decent results, just please don't expect it to be on par with the optimized versions (yet?)